refactor src.vex.time submodule
