+++ src/vex/time/__init__.py
@@ -2,27 +2,16 @@
 
 
 from collections.abc import Sequence
-from enum import IntEnum
 
 from abm.decor import act
 
 # pylint: disable=unused-import
-from .util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
+from ..util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
+
+from .time_units import TimeUnits
 
 
 __all__: Sequence[str] = 'TimeUnits', 'wait'
-
-
-@robotmesh_doc("""
-    The measurement units for time values.
-
-    robotmesh.com/studio/content/docs/vexiq-python_b/html/classvex_1_1_time_units.html
-""")
-class TimeUnits(IntEnum):
-    """Time Units."""
-
-    SEC: int = 0   # A time unit that is measured in seconds.
-    MSEC: int = 1   # A time unit that is measured in milliseconds.
 
 
 # pylint: disable=unused-argument
